requirements.txt
# 멜 스펙트로그램 실습 의존성

# 오디오 처리
librosa>=0.10.0
soundfile>=0.12.0

# 수치 연산
numpy>=1.24.0
scipy>=1.10.0

# 시각화
matplotlib>=3.7.0

# 설정 파일
pyyaml>=6.0
